+++ tekore/_sender/__init__.py
@@ -1,13 +1,4 @@
-from .concrete import (
-    Sender,
-    SyncSender,
-    AsyncSender,
-    TransientSender,
-    AsyncTransientSender,
-    PersistentSender,
-    AsyncPersistentSender,
-    SingletonSender,
-    AsyncSingletonSender,
-)
+from .base import Sender, Request, Response
+from .concrete import SyncSender, AsyncSender
 from .extending import ExtendingSender, RetryingSender, CachingSender
 from .client import SenderConflictWarning, Client